+++ CHANGELOG.md
@@ -6,6 +6,7 @@ We follow [Semantic Versions](https://semver.org/).
 ## Version 0.1.4 (unreleased)
 
 - Add `Jinja` templates with cute GIFs - used for the OAuth callback
+- Add `flask` to dependencies
 
 
 ## Version 0.1.3